1. ProfileData建模代码重构
1）将Profile的建模代码抽象成StandardProfileModeler，所以ProfileDataMessageHandler的代码被简化，使用StandardProfileModeler来实现画像缓存存储
2）提供IStandardProfileModelListener接口，用于扩展StandardProfileModeler的功能，允许在建模过程中的事件添加自定义代码

2. StandardMessagingBuilder buildComsumer增加groupName参数，支持消费同一个topic，但是consumer分组不同的情况。

3. MSCP隔离机制增强，严格限制限制跨feature的class初始化以及加载和卸载过程。
