+++ com.creditease.uav.agent.buildComponent/config/agent.properties
@@ -16,7 +16,7 @@ meta.nodetype=\u76D1\u63A7\u4EE3\u7406\u7A0B\u5E8F
 
 #resource components
 #resource.xxxx
-resource.messageproducer.class=com.creditease.agent.feature.common.MessageProducerResourceComponent
+resource.messageproducer.class=org.uavstack.resources.common.messaging.MessageProducerResourceComponent
 #resource.xxxx.properties, for example: <key1>=<value1>,<key2>=<value2>,...
 resource.messageproducer.properties=
 #resource.xxxx.<customized item>
@@ -27,7 +27,7 @@ resource.messageproducer.messagingthreshold=30
 resource.messageproducer.messagesizelimit=262000
 
 #System Http Invoker
-resource.httpinvoke.class=com.creditease.agent.feature.common.HttpInvokerResourceComponent
+resource.httpinvoke.class=org.uavstack.resources.common.invoker.HttpInvokerResourceComponent
 resource.httpinvoke.properties=
 resource.httpinvoke.http.client.maxperroute=20
 resource.httpinvoke.http.client.maxtotal=100
@@ -36,14 +36,14 @@ resource.httpinvoke.http.client.conntimeout=10000
 resource.httpinvoke.http.client.reqtimeout=10000
 
 #notifytestagent
-feature.monitortestagent.loader=default
+feature.monitortestagent.loader=com.creditease.uav.agent-1.0.jar
 #feature.xxx.enable should enable this feature
 feature.monitortestagent.enable=false
 feature.monitortestagent.components=com.creditease.agent.feature.MonitorAgentUT
 feature.monitortestagent.interval=1500
 
 #nodeoperagent
-feature.nodeoperagent.loader=default
+feature.nodeoperagent.loader=com.creditease.uav.agent.heartbeat.jar
 #feature.xxx.enable should enable this feature
 feature.nodeoperagent.enable=true
 feature.nodeoperagent.components=com.creditease.agent.feature.NodeOperAgent
@@ -51,7 +51,7 @@ feature.nodeoperagent.http.port=10101
 feature.nodeoperagent.checksec=false
 
 #notifyagent
-feature.notifyagent.loader=default
+feature.notifyagent.loader=com.creditease.uav.notifymanager-1.0.jar
 #feature.xxx.enable should enable this feature
 feature.notifyagent.enable=true
 #feature.xxx.components is the component classes (extends com.creditease.agent.spi.AgentFeatureComponent), separator "," for multiply classes
@@ -66,7 +66,7 @@ feature.notifyagent.frozenTime=300
 
 #heartbeatclientagent Agent Feature xxx
 #feature.xxx.loader should be a jar file path or default
-feature.hbclientagent.loader=default
+feature.hbclientagent.loader=com.creditease.uav.agent.heartbeat.jar
 #feature.xxx.enable should enable this feature
 feature.hbclientagent.enable=true
 #feature.xxx.components is the component classes (extends com.creditease.agent.spi.AgentFeatureComponent), separator "," for multiply classes
@@ -81,7 +81,7 @@ feature.hbclientagent.interval=15000
 #agent feature configuration
 #Monitor Agent Feature xxx
 #feature.xxx.loader should be a jar file path or default
-feature.monitoragent.loader=default
+feature.monitoragent.loader=com.creditease.uav.agent-1.0.jar
 #feature.xxx.enable should enable this feature
 feature.monitoragent.enable=true
 #feature.xxx.components is the component classes (extends com.creditease.agent.spi.AgentFeatureComponent), separator "," for multiply classes
@@ -126,7 +126,7 @@ feature.monitoragent.limit.cpu.timerange=120000
 
 #Log Agent Feature xxx
 #feature.xxx.loader should be a jar file path or default
-feature.logagent.loader=default
+feature.logagent.loader=com.creditease.uav.agent-1.0.jar
 #feature.xxx.enable should enable this feature
 feature.logagent.enable=true
 #feature.xxx.enable should enable the self Log collection feature
@@ -146,7 +146,7 @@ feature.logagent.MutiThread.thread.max=2
 feature.logagent.defrule.filter=
 
 #Proc Detect Agent Feature
-feature.procscan.loader=default
+feature.procscan.loader=com.creditease.uav.agent.heartbeat.jar
 feature.procscan.enable=true
 feature.procscan.components=com.creditease.agent.feature.ProcDetectAgent
 feature.procscan.properties=
@@ -156,13 +156,13 @@ feature.procscan.networkDetect.interval=15000
 feature.procscan.networkDetect.collectTime=2000
 
 # Process Watch Feature
-feature.procwatch.loader=default
+feature.procwatch.loader=com.creditease.uav.agent.heartbeat.jar
 feature.procwatch.enable=true
 feature.procwatch.interval=15000
 feature.procwatch.components=com.creditease.agent.feature.ProcWatchAgent
 
 # Data Collect Feature
-feature.collectclient.loader=default
+feature.collectclient.loader=com.creditease.uav.collect.jar
 feature.collectclient.enable=true
 feature.collectclient.components=com.creditease.uav.collect.client.CollectDataAgent
 feature.collectclient.properties=
@@ -177,6 +177,6 @@ feature.collectclient.idlecheckerinterval=5000
 
 
 # Thread Analysis Feature
-feature.threadanalysis.loader=default
+feature.threadanalysis.loader=com.creditease.uav.threadanalysis.jar
 feature.threadanalysis.enable=true
 feature.threadanalysis.components=com.creditease.uav.threadanalysis.client.ThreadAnalysisAgent
