+++ com.creditease.uav.agent/pom.xml
@@ -46,6 +46,11 @@
 		
 		<dependency>
 			<groupId>com.creditease.moniter</groupId>
+			<artifactId>org.uavstack.resources.common</artifactId>
+		</dependency>
+		
+		<dependency>
+			<groupId>com.creditease.moniter</groupId>
 			<artifactId>com.creditease.uav.collect</artifactId>
 		</dependency>
 		
@@ -60,6 +65,7 @@
 			<scope>test</scope>
 			<version>4.10</version>
 		</dependency>
+		
 
 	</dependencies>
