+++ com.creditease.uav.base/pom.xml
@@ -39,6 +39,13 @@
 			</plugin>
 			<plugin>
 				<artifactId>maven-jar-plugin</artifactId>
+				<configuration>    
+                <archive> 
+                	<manifestEntries>
+                			<Class-Path>../lib/com.creditease.uav.helper.jar ../lib/com.creditease.uav.fastjson.jar</Class-Path>
+            		</manifestEntries>
+                </archive>  
+              </configuration>    
 				<executions>
 					<execution>
 						<id>boot</id>
