+++ com.creditease.uav.healthmanager.buildComponent/config/agent.properties
@@ -14,7 +14,7 @@ meta.nodetype=\u5065\u5EB7\u7BA1\u7406\u7A0B\u5E8F
 
 #resource components
 #resource.xxxx
-resource.messageproducer.class=com.creditease.agent.feature.common.MessageProducerResourceComponent
+resource.messageproducer.class=org.uavstack.resources.common.messaging.MessageProducerResourceComponent
 #resource.xxxx.properties, for example: <key1>=<value1>,<key2>=<value2>,...
 resource.messageproducer.properties=
 #resource.xxxx.<customized item>
@@ -23,7 +23,7 @@ resource.messageproducer.messagingnameserver=127.0.0.1:9876
 resource.messageproducer.messagingthreshold=30
 
 #System Http Invoker
-resource.httpinvoke.class=com.creditease.agent.feature.common.HttpInvokerResourceComponent
+resource.httpinvoke.class=org.uavstack.resources.common.invoker.HttpInvokerResourceComponent
 resource.httpinvoke.properties=
 resource.httpinvoke.http.client.maxperroute=20
 resource.httpinvoke.http.client.maxtotal=100
@@ -33,14 +33,14 @@ resource.httpinvoke.http.client.reqtimeout=10000
 
 
 #nodeoperagent
-feature.nodeoperagent.loader=default
+feature.nodeoperagent.loader=com.creditease.uav.agent.heartbeat.jar
 #feature.xxx.enable should enable this feature
 feature.nodeoperagent.enable=true
 feature.nodeoperagent.components=com.creditease.agent.feature.NodeOperAgent
 feature.nodeoperagent.http.port=10102
 
 #notifcenter
-feature.notifycenter.loader=default
+feature.notifycenter.loader=com.creditease.uav.notifycenter.jar
 #feature.xxx.enable should enable this feature
 feature.notifycenter.enable=true
 #feature.xxx.components is the component classes (extends com.creditease.agent.spi.AgentFeatureComponent), separator "," for multiply classes
@@ -89,10 +89,10 @@ feature.notifycenter.http.core=10
 feature.notifycenter.http.max=50
 feature.notifycenter.http.bqsize=10
 #push to other services
-feature.notifycenter.push.services=
+feature.notifycenter.push.services=dhandsonrobot-DHandsonRobot-Server-/hit/xrobot:monitor/ntfpush
 
 #notifyagent
-feature.notifyagent.loader=default
+feature.notifyagent.loader=com.creditease.uav.notifymanager.jar
 #feature.xxx.enable should enable this feature
 feature.notifyagent.enable=true
 #feature.xxx.components is the component classes (extends com.creditease.agent.spi.AgentFeatureComponent), separator "," for multiply classes
@@ -108,7 +108,7 @@ feature.notifyagent.frozenTime=300
 #agent feature configuration
 #Monitor Agent Feature xxx
 #feature.xxx.loader should be a jar file path or default
-feature.healthmanager.loader=default
+feature.healthmanager.loader=com.creditease.uav.healthmanager-1.0.jar
 #feature.xxx.enable should enable this feature
 feature.healthmanager.enable=true
 #feature.xxx.components is the component classes (extends com.creditease.agent.spi.AgentFeatureComponent), separator "," for multiply classes
@@ -204,7 +204,7 @@ feature.healthmanager.MT_Log.ds.pagesize=3000
 
 #heartbeatserveragent Agent Feature xxx
 #feature.xxx.loader should be a jar file path or default
-feature.hbserveragent.loader=default
+feature.hbserveragent.loader=com.creditease.uav.agent.heartbeat.jar
 #feature.xxx.enable should enable this feature
 feature.hbserveragent.enable=true
 #feature.xxx.components is the component classes (extends com.creditease.agent.spi.AgentFeatureComponent), separator "," for multiply classes
@@ -239,7 +239,7 @@ feature.hbserveragent.ntp.enable=false
 
 #heartbeatclientagent Agent Feature xxx
 #feature.xxx.loader should be a jar file path or default
-feature.hbclientagent.loader=default
+feature.hbclientagent.loader=com.creditease.uav.agent.heartbeat.jar
 #feature.xxx.enable should enable this feature
 feature.hbclientagent.enable=true
 #feature.xxx.components is the component classes (extends com.creditease.agent.spi.AgentFeatureComponent), separator "," for multiply classes
@@ -249,10 +249,11 @@ feature.hbclientagent.properties=
 #feature.monitoragent.hbhandlers for heartbeat handlers
 feature.hbclientagent.hbhandlers=com.creditease.agent.feature.hbagent.handlers.HBClientDefaultHandler
 feature.hbclientagent.hbservers=127.0.0.1:8010
+
 feature.hbclientagent.interval=15000
 
 #Proc Detect Agent Feature
-feature.procscan.loader=default
+feature.procscan.loader=com.creditease.uav.agent.heartbeat.jar
 feature.procscan.enable=false
 feature.procscan.components=com.creditease.agent.feature.ProcDetectAgent
 feature.procscan.properties=
@@ -260,7 +261,7 @@ feature.procscan.properties=
 feature.procscan.scan.interval=15000
 
 #feature apphubmanager  BEGIN
-feature.apphubmanager.loader=default
+feature.apphubmanager.loader=com.creditease.uav.healthmanager-1.0.jar
 feature.apphubmanager.enable=true
 feature.apphubmanager.components=com.creditease.uav.feature.AppHubManager
 feature.apphubmanager.appHubManagerHandlers=com.creditease.uav.feature.apphubmanager.http.handlers.AHDataStoreAppHandler,com.creditease.uav.feature.apphubmanager.http.handlers.AHDataStoreGroupHandler,com.creditease.uav.feature.apphubmanager.http.handlers.AHDataStoreFeedbackHandler
@@ -279,7 +280,7 @@ feature.apphubmanager.ds.pwd=
 #feature apphubmanager  END
 
 #IssueDiagnoseAssitAgent
-feature.ida.loader=default
+feature.ida.loader=com.creditease.uav.healthmanager-1.0.jar
 #feature.xxx.enable should enable this feature
 feature.ida.enable=false
 #feature.xxx.components is the component classes (extends com.creditease.agent.spi.AgentFeatureComponent), separator "," for multiply classes
@@ -293,7 +294,7 @@ feature.ida.ContainerPerfCatcher.threshold=cpu=0.8
 feature.ida.ContainerPerfCatcher.root=d:/UAVTEST
 
 #feature runtimenotify 
-feature.runtimenotify.loader=default
+feature.runtimenotify.loader=com.creditease.uav.healthmanager-1.0.jar
 feature.runtimenotify.enable=true
 feature.runtimenotify.components=com.creditease.uav.feature.RuntimeNotifyCatcher
 feature.runtimenotify.rnhandlers=com.creditease.uav.feature.runtimenotify.RuntimeNotifyDataHandler
@@ -339,7 +340,7 @@ feature.runtimenotify.strategy.interval=30000
 feature.runtimenotify.slice.config={"N":{"timerange":"600000","expire":"60000"},"M":{"timerange":"600000","expire":"60000"},"L":{"timerange":"60000","expire":"60000"}}
 
 # Data Collect Feature
-feature.collectserver.loader=default
+feature.collectserver.loader=com.creditease.uav.collect.jar
 feature.collectserver.enable=true
 feature.collectserver.components=com.creditease.uav.collect.server.GatherDataManager
 feature.collectserver.properties=
@@ -366,7 +367,7 @@ feature.collectserver.JQ_JTA.consumestopinterval=0
 feature.collectserver.JQ_JTA.enable=true
 
 #InvokeChain Process Server Feature
-feature.invokechainserver.loader=default
+feature.invokechainserver.loader=com.creditease.uav.invokechain.jar
 feature.invokechainserver.enable=true
 feature.invokechainserver.components=com.creditease.uav.invokechain.InvokeChainProcessServer
 feature.invokechainserver.properties=
@@ -391,7 +392,7 @@ feature.invokechainserver.qhandlers=com.creditease.uav.invokechain.http.InvokeCh
 
 
 #HealthManager NewLogService Feature
-feature.newlogservice.loader=default
+feature.newlogservice.loader=com.creditease.uav.invokechain.jar
 feature.newlogservice.enable=true
 feature.newlogservice.components=com.creditease.uav.healthmanager.newlog.HMNewLogService
 feature.newlogservice.properties=
@@ -417,7 +418,7 @@ feature.newlogservice.qhandlers=com.creditease.uav.healthmanager.newlog.http.New
 
 
 #Thread Analysis Server Feature
-feature.threadanalysisserver.loader=default
+feature.threadanalysisserver.loader=com.creditease.uav.threadanalysis.jar
 feature.threadanalysisserver.enable=true
 feature.threadanalysisserver.components=com.creditease.uav.threadanalysis.server.ThreadAnalysisServer
 #elastic-search
