+++ com.creditease.uav.invokechain/pom.xml
@@ -36,7 +36,11 @@
 			<artifactId>com.creditease.uav.cache.redis</artifactId>
 
 		</dependency>
-
+		
+		<dependency>
+			<groupId>com.creditease.moniter</groupId>
+			<artifactId>org.uavstack.resources.common</artifactId>
+		</dependency>
 		<!-- <dependency>
 			<groupId>org.elasticsearch.client</groupId>
 			<artifactId>transport</artifactId>
