+++ com.creditease.uav.messaging/src/main/java/com/creditease/uav/messaging/api/MessageConsumer.java
@@ -20,7 +20,6 @@
 
 package com.creditease.uav.messaging.api;
 
-
 public abstract class MessageConsumer {
 
     protected String[] msgTypeIds;
