+++ com.creditease.uav.monitorframework.apm/findbugs.exclude.xml
@@ -13,6 +13,10 @@
     	<Bug pattern="NP_ALWAYS_NULL" />
    </Match>
    <Match>
+     <Class name="com.creditease.uav.apm.invokechain.handlers.ClientSpanInvokeChainHandler" />  
+    	<Bug pattern="RCN_REDUNDANT_NULLCHECK_OF_NONNULL_VALUE" />
+   </Match>
+   <Match>
 	   		<Bug pattern="DM_DEFAULT_ENCODING" />
 	   </Match>
 </FindBugsFilter>
