+++ com.creditease.uav.upgrade/config/upgrade.properties
@@ -14,7 +14,7 @@ meta.nodetype=\u5347\u7ea7\u7a0b\u5e8f
 
 #heartbeatclientagent Agent Feature xxx
 #feature.xxx.loader should be a jar file path or default
-feature.hbclientagent.loader=default
+feature.hbclientagent.loader=com.creditease.uav.agent.heartbeat.jar
 #feature.xxx.enable should enable this feature
 feature.hbclientagent.enable=true
 #feature.xxx.components is the component classes (extends com.creditease.agent.spi.AgentFeatureComponent), separator "," for multiply classes
@@ -27,7 +27,7 @@ feature.hbclientagent.hbservers=127.0.0.1:8010
 feature.hbclientagent.interval=15000
 
 #nodeoperagent
-feature.nodeoperagent.loader=default
+feature.nodeoperagent.loader=com.creditease.uav.agent.heartbeat.jar
 #feature.xxx.enable should enable this feature
 feature.nodeoperagent.enable=true
 feature.nodeoperagent.components=com.creditease.agent.feature.NodeOperAgent
