+++ org.uavstack.resources.common/.gitignore
@@ -1,0 +1,2 @@
+/build/
+/target/
