+++ org.uavstack.resources.common/findbugs.exclude.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0"?>
+<FindBugsFilter>  
+  
+</FindBugsFilter>
