+++ org.uavstack.resources.common/pom.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
 
 	<parent>
 		<artifactId>ce-datamonitorsystem</artifactId>
@@ -9,53 +8,53 @@
 		<version>1.0</version>
 		<relativePath>../com.creditease.uav.superpom</relativePath>
 	</parent>
+	<modelVersion>4.0.0</modelVersion>
 
-	<artifactId>com.creditease.uav.collect</artifactId>
+	<artifactId>org.uavstack.resources.common</artifactId>
 
 	<dependencies>
 		<dependency>
 			<groupId>com.creditease.moniter</groupId>
 			<artifactId>com.creditease.uav.helper</artifactId>
-		</dependency>
 
+		</dependency>
 		<dependency>
 			<groupId>com.creditease.moniter</groupId>
 			<artifactId>com.creditease.uav.base</artifactId>
-		</dependency>
 
+		</dependency>
+		
+		<dependency>
+			<groupId>com.creditease.moniter</groupId>
+			<artifactId>com.creditease.uav.httpasync</artifactId>
+
+		</dependency>
+		
+		<dependency>
+			<groupId>com.creditease.moniter</groupId>
+			<artifactId>com.creditease.uav.cache.redis</artifactId>
+
+		</dependency>
+		
 		<dependency>
 			<groupId>com.creditease.moniter</groupId>
 			<artifactId>com.creditease.uav.messaging</artifactId>
 		</dependency>
-		<dependency>
-			<groupId>com.creditease.moniter</groupId>
-			<artifactId>com.creditease.uav.notifymanager</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>com.creditease.moniter</groupId>
-			<artifactId>org.uavstack.resources.common</artifactId>
-		</dependency>
 		
-		<!-- @Deprecated -->
-		<dependency>
-			<groupId>com.google.guava</groupId>
-			<artifactId>guava</artifactId>
-			<version>11.0.2</version>
-		</dependency>
-
-	</dependencies>
-
+		
+	</dependencies>	
+	
 	<build>
-		<plugins>
-			<plugin>
+        <plugins>
+         
+            <plugin>
 				<artifactId>maven-compiler-plugin</artifactId>
-				<!-- <version>3.3</version> -->
 				<configuration>
 					<source>1.7</source>
 					<target>1.7</target>
 					<encoding>UTF-8</encoding>
 				</configuration>
 			</plugin>
-		</plugins>
-	</build>
+        </plugins>
+    </build>
 </project>
