+++ org.uavstack.resources.common/src/main/java/org/uavstack/resources/common/invoker/HttpInvokerResourceComponent.java
@@ -18,7 +18,7 @@
  * >>
  */
 
-package com.creditease.agent.feature.common;
+package org.uavstack.resources.common.invoker;
 
 import com.creditease.agent.spi.AbstractSystemInvoker;
 import com.creditease.agent.spi.AgentResourceComponent;
